+++ common-tuple/change-log.md
@@ -1,1 +1,3 @@
-# 0.1.0 2023
+# 1.0.0 2024-10-24
+1. fork from speedment tuple
+2. MutableTuple添加set(index,Object)方法